+++ python/GafferUITest/StandardNodeUITest.py
@@ -57,22 +57,5 @@ class StandardNodeUITest( GafferUITest.TestCase ) :
 		self.assertTrue( isinstance( u.plugValueWidget( n["c"]["i"] ), GafferUI.PlugValueWidget ) )
 		self.assertTrue( u.plugValueWidget( n["c"]["i"] ).getPlug().isSame( n["c"]["i"] ) )
 
-	def testSetReadOnlyForUserPlugs( self ) :
-
-		n = Gaffer.Node()
-		n["user"]["a"] = Gaffer.IntPlug( flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
-
-		u = GafferUI.StandardNodeUI( n )
-		self.assertEqual( u.plugValueWidget( n["user"]["a"] ).getReadOnly(), False )
-
-		u.setReadOnly( True )
-		self.assertEqual( u.plugValueWidget( n["user"]["a"] ).getReadOnly(), True )
-
-		u = GafferUI.StandardNodeUI( n )
-		w = u.plugValueWidget( n["user"]["a"] )
-
-		u.setReadOnly( True )
-		self.assertEqual( w.getReadOnly(), True )
-
 if __name__ == "__main__":
 	unittest.main()